Temp fix for current dev
This is a temporary fix for the current dev versions, above 3.0.175 (I believe tags were added in 3.0.176). This will prevent the JS errors by disabling the module from trying to initialize the tags field, while keeping the module installed and retaining your settings, as well as the ability to use selectize on the icons.